Refactor RGBMatrix to control the hardware pixels and introduce Engine to manage animations.
